+++ integrationtest/ValidationProject/ValidateStrykerResults.cs
@@ -144,20 +144,6 @@ namespace IntegrationTests
             CheckReportTestCounts(report, total: 30);
         }
 
-        [Fact]
-        [Trait("Category", "InitCommand")]
-        public void InitCommand()
-        {
-            var directory = new DirectoryInfo("../../../../TargetProjects/InitCommand");
-            var jsonFile = directory.GetFiles("*.json", SearchOption.AllDirectories).SingleOrDefault();
-            jsonFile.ShouldNotBeNull("Json file missing");
-
-            var configOutput = File.ReadAllText(jsonFile.FullName);
-            var parsedConfigOuptut = JsonConvert.DeserializeObject<FileBasedInputOuter>(configOutput);
-
-            parsedConfigOuptut.Input.Project.ShouldBe("TestProject.csproj");
-        }
-
         private void CheckMutationKindsValidity(JsonReport report)
         {
             foreach (var file in report.Files)
