+++ src/Stryker.CLI/Stryker.CLI.UnitTest/CommandLineConfigReaderTests.cs
@@ -14,7 +14,7 @@ namespace Stryker.CLI.UnitTest
             Name = "Stryker",
             FullName = "Stryker: Stryker mutator for .Net",
             Description = "Stryker mutator for .Net",
-            ExtendedHelpText = "Welcome to Stryker for .Net! Run dotnet stryker to kick off a mutation test run"
+            ExtendedHelpText = "Welcome to Stryker for .Net! Run dotnet stryker to kick off a mutation test run or run dotnet stryker init to start configuring your project."
         };
         private readonly IStrykerInputs _inputs = new StrykerInputs();
         private readonly CommandLineConfigReader _target = new();
